Gopkg.toml: Set seamia/tools to version v1.1.8, not master branch
Switches to specifying a dependency to a specific git tag [1] rather than the
master branch, for a dependency which happens to be another repository of the
protodot author.

Note: the author of this specific commit is not currently familiar with go.
When trying to update the "Gopkg.lock" file it became apparent the tool is used
by Go's *deprecated* "dep" dependency management tool rather than the modern
"go mod" tool which uses "go.mod" file (which is present in this repository).
This commit still will be useful for systems which haven't been updated though.

By pointing to a specific git tag, users building locally will no longer
unknowingly pull in newer (in future, possibly API breaking) changes from that
repository, as new commits are inevitably added to its master branch.

The most recent release of protodot is v1.2.0 (2023-04-12), which based on
timestamps must have originally been built with seamia/tools v1.0.12
(2021-08-08). However, the most recent binary upload (hosted within the
repository) was commited on 2023-12-12, which indicates it was likely built
with seamia/tools v1.1.8 (2023-11-30)

[1] https://github.com/seamia/tools/tree/v1.1.8